Fixed map layer ordering issue
Desired map behavior includes geo-search selection highlight layer appearing above the collection-hover highlight layer. Previously, the hover layer would throw an error in the console when a geo-search layer was not drawn as there was no search layer to stack underneath. Now, the hover layer only tries to stack beneath the geosearch layer if it is set in recoil atom.