changed endpoint for catalog cards to pull from new unified schema endpt
KNOWN ISSUE: 

new naip collection is viewable in catalog view but metadata does not show in collection view. This is because the /collection endpoint is pulling from the non-materialized view and needs to be re-created as a legitimate materialized view that pulls and refreshed from the collection_catalog_record table and the collection_coverage_geom tables.